+++ scripts/sd_peft.py
@@ -508,7 +508,8 @@ def train(cfg: SDPeftConfig, strategies: "SdPeftStrategy"):
     if is_main_process:
         timestep_counts, plotter_settings = setup_live_plotter(cfg, noise_scheduler, strategies.la_sampler, strategies)
 
-    edm2_model, edm2_optimizer, edm2_lr_scheduler = prepare_edm2_loss_weighting(cfg.loss, cfg.training, noise_scheduler, accelerator)
+    edm2_model, edm2_optimizer, edm2_lr_scheduler = prepare_edm2_loss_weighting(cfg.loss, cfg.training, noise_scheduler,
+                                                                                accelerator)
 
     init_trackers(accelerator, cfg.output.logging, "adapter_train")
 
@@ -609,7 +610,8 @@ def train(cfg: SDPeftConfig, strategies: "SdPeftStrategy"):
         accelerator.unwrap_model(adapter).train()
 
     if plot_edm2_loss_weighting_check(cfg.loss, cfg.training, global_step):
-        plot_edm2_loss_weighting(cfg.loss, cfg.output.saving.output_name, global_step, edm2_model, 1000, accelerator.device)
+        plot_edm2_loss_weighting(cfg.loss, cfg.output.saving.output_name, global_step, edm2_model, 1000,
+                                 accelerator.device)
 
     is_tracking = len(accelerator.trackers) > 0
     if is_tracking:
@@ -841,7 +843,8 @@ def train(cfg: SDPeftConfig, strategies: "SdPeftStrategy"):
                                     remove_model(remove_loss_weights_ckpt_name)
 
                     if plot_edm2_loss_weighting_check(cfg.loss, cfg.training, global_step):
-                        plot_edm2_loss_weighting(cfg.loss, cfg.output.saving.output_name, global_step, edm2_model, 1000, accelerator.device)
+                        plot_edm2_loss_weighting(cfg.loss, cfg.output.saving.output_name, global_step, edm2_model, 1000,
+                                                 accelerator.device)
                     optimizer_train_fn()
                     accelerator.unwrap_model(adapter).train()